Block requests to `xmlrpc.php`
This restriction blocks requests to `xmlrpc.php`, which in turn will
hopefully mitigate some server problems we've been experiencing.
According to...

  https://www.digitalocean.com/community/tutorials/
    how-to-protect-wordpress-from-xml-rpc-attacks-on-ubuntu-14-04

...it's possible that we've been seeing some brute-force attacks in
which requests are made for `xmlrpc.php` until the server crashes. Since
none of our tools or services currently rely on this file, blocking
access to it is the easiest way to mitigate against the issue.